Update to allow fractional timeoffset values for the shape and erma_data_package outputters.

See merge request gnome/pygnome!147

+++ py_gnome/gnome/outputters/erma_data_package.py
@@ -120,7 +120,7 @@ class ERMADataPackageSchema(BaseOutputterSchema):
     # Currently we do not support timezone, but instead stick with static
     # time offset.
     # timezone = SchemaNode(String(), save=True, update=True)
-    timeoffset = SchemaNode(Int(), save=True, update=True)
+    timeoffset = SchemaNode(Float(), save=True, update=True)
     enable_each_timestep_as_layer = SchemaNode(
         Boolean(), save=True, update=True
     )
